Debug packets can now be viewed by specifying #define PARTICLE_DEBUG
When objects didn't get their expected slot, bad packets were being transmitted.  This was because ACLSendPacket() was called when it wanted to transmit, and then when it wasn't able to (because another object did), it went on recording readings and putting them on the radio stack using ACLSendData().

It has been changed now to not take any readings if it is waiting to transmit.  This is not ideal, since any readings it could be taking while waiting to transmit are lost.