Add first shot of automated Delphi-Unit build job. This yet a bit bleeding on the edges, but builds a working Delphi-Unit at the moment.
Some refining the awks and sed and this will be a C2Delphi parser ;)

It messes up the sytle in the .h files a bit in that it requires:
 - ^DLL lines to *not* wrap
 - wants the ";" right after the closing ")"
 - knows only about the currently used types
 - replaces only a limited number of types

The Makefile needs a automake/autoconf round, I am afraid. Zero knowledge of
it here :-\ ... I guess/hope ${GREP}, ${SED} and ${AWK} are available in any
POSIX build environment?!? We can still ship with a prebuild unit, though.