Disabled Exception Handling.
Since one can't bind the exception handler of SWIG to specific return type, the
solution was to check the p_errno != 0 after every call and reset it afterwards.
But there can occur Conditions where the p_errno variable is set != 0 and no
Error occured. For Example when an error occured but got handled in the
libparticle without resetting the p_errno afterwards.

possible solutions:
- make sure libparticle or wrapper calls reset the p_errno variable
- different kind of exception handling in libparticle
- bind the exception handler to every function which needs it, this means
  writing 3 extra lines per function.