tutorial added, packetDelivery Mechanism reworked, Console completed.
Done a BeanBuilder Tutorial, Eclipse Tutorial to come. Currently integrated with
the api documentation. To read it you have to have doxygen generate the Docs.

Packets aren't anymore directly delivered to Components but are requested by
them. A ParticleEvent is instead fired on Arrival of new Packets. The
SocketThread to get the Packets from is a Property of the Component.

The Console is feature complete:
- scrolling from newest to oldest packets. 
- alternate highlighting.