I might get killed for this during the next days:
moved "lib" to "libparticle"
this is nessesary !!!!
if we want to install libparticle on some system, we cannot write
something like conf.h into the main include path.
furthermore #include <libparticle.h> should suffice if you write some app

If I don't it now, it will be much more of a pain later.