firmware v1.20
Lack of synchronisation on the Awarecon layer was having bad side
effects, the most observable being a decrease in measurement rate.
This was especially significant when there was no non-Relate particle
(such as an Xbridge) running in the vicinity of the Relate devices.
The exact reason for the lack of sync is unknown, but it could have
something to do with the fact that Relate devices are in
RF_SYNC_MODE_BLIND_FLIGHT so often.  However, theoretically this
shouldn't affect it, since sync is still performed even in blind
flight mode.

Regardless, a master/slave sync strategy solves this, allowing higher
measurement rates and more consistent performance than observed with
the default Awarecon stack.  The Relate device with the highest ID is
the master which sends out the sync, and all other Relate devices are
slaves which only listen for the sync.  If the master drops out, a new
one takes over after a specified amount of time (currently 8 seconds).
To utilise master/slave operation, #define
RELATE_SYNC_MODE_MASTER_SLAVE.

Other changes:

- re-measured RTC1HOUR_RESPONSE_LAG, and provided comments that better
explain how it is measured

- removed some old/debugging code which was commented out