firmware v1.26
Added functionality for Particle v2.29.  Note that the Particle 2.29
boards do not include pullups on some of the lines that are used for
I2C communication to the temperature sensor on Relate boards.  So, any
Relate board used with Particle 2.29 must include pullup resistors on
the Conan pins used for I2C.  Currently, only the DIY_ADAPTOR has
pullups.