firmware v1.31
Improved the way USBReceive() reads in and echoes back data.  Leading
junk bytes of 0xAA from the host are ignored.  Dongle also sends
trailing bytes of 0xFF after it echoes back what the host sent.  (This
is a workaround to counteract the fact that the last bytes of the echo
were sometimes corrupted for unknown reasons.  Now, the 0xFF bytes are
corrupted instead.)

If the dongle is interrupted by the host, authentication stops.  Of
course, the host can then send a command string to the dongle to tell
it to start a new authentication sequence.