Avoid Looping if a UPnP Specification doesn't match a packet for receiving but just matches for sending to this particle.
Example:

APS specifies that crs packet can be send to reprogram this particle, but it doesn't match on packets sent by this particle as the spec doesn't subscribe to sgx, sgy ... 
It happened that a Template with matches to only crs was created, but not on the aps packets sent by the particle. 

Solution:
IN Arguments Tuple Types to the particle are not added to the Template.