fixed c/c++ problems with struct soap and implemented multicast
In gSOAP it is currently not possible to mix c and c++ code. The struct soap
member alignment is different in gcc and g++ in that way that g++ adds a
pointer to the vtable at the beginning. That is the reason why gsoap-nb-io.cpp
has to be converted to plain c and a c++ wrapper is implemented in
cpp_wrap.cpp.

The NetBurner library does not support to use select() in conjunction with
multicast. The mymulticast.cpp file is a copy of the original NetBurner
multicast.cpp file with modification to add the necessary notification
mechanism. A documentation can be found in the NetBurner forum:
http://forum.embeddedethernet.com/viewtopic.php?p=2548#p2548