initial NetBurner tutorial code included
The NetBurner airconditioner code is based on the standard simple
airconditioner code with two modifications. First of all the NetBurner I/O is
activated. Additionally eventing is implemented within the main loop avoiding
the need for threading.

For compilation reason the main function was renamed to main2 and is called
from NetBurner UserMain() function in nbmain.cpp after initializing the board.