Initial plug and play support for NetBurner boards.
When detaching the board from Ethernet and reattaching it in an other subnet
the stack is shutdown and started with the new IP address obtained with DHCP.
This way it is possible to achieve plug and play in the first case. When
reattaching to the same subnet and obtaining the same IP address nothing
happens to the stack. It is assumed that just packets get lost. In a further
modification maybe a dpws_rebind is implemented to correctly rebind with
correct transmission of bye and hello message. Since the currently shutdown is
started when the new IP is assigned a bye message is send before the hello
message.